remove MemBlockSub.scala
